+++ main.tf
@@ -17,6 +17,13 @@ module "meshcloud_account_metering_access" {
   privileged_external_id               = var.cost_explorer_privileged_external_id
   management_account_service_role_name = var.cost_explorer_management_account_service_role_name
   meshcloud_account_service_user_name  = var.cost_explorer_meshcloud_account_service_user_name
+
+  workload_identity_federation = var.workload_identity_federation == null ? null : {
+    issuer                = var.workload_identity_federation.issuer,
+    audience              = var.workload_identity_federation.audience,
+    subject               = var.workload_identity_federation.kraken_subject,
+    identity_provider_arn = aws_iam_openid_connect_provider.meshstack[0].arn
+  }
 }
 
 module "meshcloud_account_replicator_access" {
@@ -30,6 +37,13 @@ module "meshcloud_account_replicator_access" {
   meshcloud_account_service_user_name  = var.meshcloud_account_service_user_name
   management_account_service_role_name = var.management_account_service_role_name
   automation_account_service_role_name = var.automation_account_service_role_name
+
+  workload_identity_federation = var.workload_identity_federation == null ? null : {
+    issuer                = var.workload_identity_federation.issuer,
+    audience              = var.workload_identity_federation.audience,
+    subject               = var.workload_identity_federation.replicator_subject,
+    identity_provider_arn = aws_iam_openid_connect_provider.meshstack[0].arn
+  }
 }
 
 module "management_account_metering_access" {
@@ -41,6 +55,8 @@ module "management_account_metering_access" {
   privileged_external_id               = var.cost_explorer_privileged_external_id
   management_account_service_role_name = var.cost_explorer_management_account_service_role_name
   meshcloud_account_service_user_name  = var.cost_explorer_meshcloud_account_service_user_name
+
+  allow_federated_role = var.workload_identity_federation != null
 
   depends_on = [
     module.meshcloud_account_metering_access
@@ -62,6 +78,8 @@ module "management_account_replicator_access" {
   management_account_service_role_name = var.management_account_service_role_name
   landing_zone_ou_arns                 = var.landing_zone_ou_arns
 
+  allow_federated_role = var.workload_identity_federation != null
+
   depends_on = [
     module.meshcloud_account_replicator_access
   ]
@@ -76,6 +94,8 @@ module "automation_account_replicator_access" {
   privileged_external_id               = var.replicator_privileged_external_id
   meshcloud_account_service_user_name  = var.meshcloud_account_service_user_name
   automation_account_service_role_name = var.automation_account_service_role_name
+
+  allow_federated_role = var.workload_identity_federation != null
 
   depends_on = [
     module.meshcloud_account_replicator_access