+++ modules/meshcloud-replicator/replicator-automation-account-access/data.tf
@@ -10,10 +10,21 @@ data "aws_iam_policy_document" "meshfed_automation_assume_role" {
   version = "2012-10-17"
   statement {
     effect = "Allow"
+
     principals {
       type        = "AWS"
       identifiers = ["arn:${data.aws_partition.current.partition}:iam::${var.meshcloud_account_id}:user/${var.meshcloud_account_service_user_name}"]
     }
+
+    dynamic "principals" {
+      for_each = var.allow_federated_role ? [true] : []
+
+      content {
+        type        = "AWS"
+        identifiers = ["arn:${data.aws_partition.current.partition}:iam::${var.meshcloud_account_id}:role/${var.meshcloud_account_service_user_name}IdentityFederation"]
+      }
+    }
+
     actions = ["sts:AssumeRole"]
     condition {
       test     = "StringEquals"