+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added ARNs of managed accounts roles to output
 - Added meshStack access role to output
 - Renamed metering related parts from kraken to metering
+- Added workload identity federation
+- Added option to disable access keys
 
 ## [v0.1.0]